==> Analysis_sc_R/03_scRNA_BCRseq_cluster_dobletes-nf.Rmd <==
---
title: "Analysis_all_mouse"
author: "Enric Vercher"
date: "2023-10-18"
output: html_document
editor_options: 
  chunk_output_type: inline
---

```{r}
suppressPackageStartupMessages({library(parallel)
library(Seurat)
library(ggplot2)
library(dplyr)
library(tidyverse)
library(patchwork)
library(DoubletFinder)
library(sctransform)
library(glmGamPoi)
library(SeuratDisk)
library(SeuratData)
library(SeuratWrappers)
library(Nebulosa)
library(scales)
library(scRepertoire)
library(knitr)
library(UCell)
library(Nebulosa)
library(knitr)
library(SoupX)
library(DropletUtils) 
library(UCell)
#para las paletas de colores personalizadas
library(scCustomize)
library(qs)})
suppressPackageStartupMessages(library(ComplexHeatmap))
```

# DESPUÉS DE UNIRLO CON EL BCR:

```{r}
data.RNA.BCR <- readRDS(file="C:/Users/d940401/Desktop/Analisis_immcantation/Analisis_Enric/Datos_Enric/Datos_TFM/00_Datos_filtrados_slurm/07_Analisis_airr_nextflow/scRNA_con_BCR_nextflow.rds")
```

```{r}
table(data.RNA.BCR@meta.data$contains_bcr)
```

```{r}
table(data.RNA.BCR$orig.ident)
```

Sin quitar el BCR:

```{r}
data.mice.total <- data.RNA.BCR
dim(data.mice.total)
```

# QUITAMOS LOS GENES QUE NO TIENEN BCR

```{r}
# Filtrar el objeto Seurat
data.filtrado.BCR <- subset(data.RNA.BCR, subset = contains_bcr == TRUE)
# Verificar el número de células después del filtrado
dim(data.filtrado.BCR)
```

Ahora tenemos 6085 LB que sabemos que presentan un BCR funcional y adecuado.

# quitamos o no el cell cycle?

```{r}
#esto a mano, te lo pasa a minuscula, funciona muy bien
m.s.genes <- str_to_title(cc.genes.updated.2019$s.genes)
m.g2m.genes <- str_to_title(cc.genes.updated.2019$g2m.genes)
```

```{r}
# Cell-cycle score and regression
data.RNA.BCR.cellcyc <- CellCycleScoring(data.filtrado.BCR, s.features = m.s.genes, g2m.features = m.g2m.genes,set.ident = TRUE)
# Identify the most variable genes
data.RNA.BCR.cellcyc <- FindVariableFeatures(data.RNA.BCR.cellcyc, 
                     selection.method = "vst",
                     nfeatures = 2000, 
                     verbose = FALSE)
		     
# Scale the counts
data.RNA.BCR.cellcyc <- ScaleData(data.RNA.BCR.cellcyc)
# Perform PCA
data.RNA.BCR.cellcyc <- RunPCA(data.RNA.BCR.cellcyc)

# Plot the PCA colored by cell cycle phase
DimPlot(data.RNA.BCR.cellcyc,
        reduction = "pca",
        group.by= "Phase",
        split.by = "Phase")
```

Las células no se separan por ciclo celular. No es una fuente de variabilidad importante en estos datos.

# integración todas las células sólo CON BCR

Creamos la funcion integrar de CCA:

fuente: https://github.com/quadbio/scRNAseq_analysis_vignette/blob/master/Tutorial.md
fuente: https://satijalab.org/seurat/articles/integration_rpca

Mirar este paper para explicar esta parte de la integración en métodos:

https://doi.org/10.1016/j.cell.2019.05.031

"Comprehensive Integration of Single-Cell Data"

Creamos una lista con todos los ratones. Normalizamos los datos para aplicar el Cellcyclescoring

```{r}
mice.list.BCR <- SplitObject(data.filtrado.BCR, split.by = "orig.ident")
# Loop over the list of Seurat objects
# Loop sobre la lista de objetos Seurat
for (i in 1:length(mice.list.BCR)) {
  # Normalizar los datos
  mice.list.BCR[[i]] <- NormalizeData(mice.list.BCR[[i]], normalization.method = "LogNormalize", scale.factor = 10000)
  
  # Realizar el scoring del ciclo celular
  # Por favor, reemplaza 'm.g2m.genes' y 'm.s.genes' con tus conjuntos de genes reales
  mice.list.BCR[[i]] <- CellCycleScoring(mice.list.BCR[[i]], g2m.features = m.g2m.genes, s.features = m.s.genes)
  
  # Ejecutar SCTransform
  mice.list.BCR[[i]] <- SCTransform(mice.list.BCR[[i]], method = "glmGamPoi", vars.to.regress = c("percent.mt", "percent_ribo"), vst.flavor = "v2", verbose = FALSE, variable.features.n = 3000)
  
  # Ejecutar PCA
  mice.list.BCR[[i]] <- RunPCA(mice.list.BCR[[i]])
  
  # Ejecutar UMAP
  mice.list.BCR[[i]] <- RunUMAP(mice.list.BCR[[i]], dims = 1:20)
}

```

```{r}
# Loop over the list of Seurat objects
for (i in 1:length(mice.list.BCR)) {
  # Identify the 15 most highly variable genes
  top_genes <- VariableFeatures(mice.list.BCR[[i]], assay = "SCT")[1:15]
  
  # Plot the average expression and variance of these genes
  p <- LabelPoints(plot = VariableFeaturePlot(mice.list.BCR[[i]]), points = top_genes, repel = TRUE)
  
  # Print the plot (optional)
  print(p)
}
```

# IDENTIFICACIÓN DE LOS DOBLETES en cada ratón individual

Se aplica sobre cada ratón individual.

```{r}
# Crear una lista para almacenar los resultados de los dobletes de cada ratón
df_results_list <- list()

# Loop sobre la lista de objetos Seurat
for (i in 1:length(mice.list.BCR)) {
  # Identificar el número óptimo de componentes principales (PCs)
  m.sweep.res <- paramSweep(mice.list.BCR[[i]], PCs = 1:20, sct = TRUE)
  sweep.stats <- summarizeSweep(m.sweep.res)
  bcmvn <- find.pK(sweep.stats)
  optimal.pk <- as.numeric(levels(bcmvn$pK))[bcmvn$pK][which.max(bcmvn$BCmetric)]
  
  # Calcular la proporción de doublets homotípicos
  annotations <- mice.list.BCR[[i]]@meta.data$seurat_clusters
  homotypic.prop <- modelHomotypic(annotations)           
  nExp_poi <- round(optimal.pk * nrow(mice.list.BCR[[i]]@meta.data))
  nExp_poi.adj <- round(nExp_poi * (1 - homotypic.prop))
  
  # Ejecutar DoubletFinder
  df_result <- doubletFinder(seu = mice.list.BCR[[i]], PCs = 1:20, pK = optimal.pk, nExp = nExp_poi.adj, sct = TRUE)
  
  # Obtener el nombre de la columna de la predicción de DoubletFinder
  DF.name <- colnames(df_result@meta.data)[grepl("DF.classification", colnames(df_result@meta.data))]
  
  # Mostrar los gráficos si es necesario
  cowplot::plot_grid(ncol = 2, 
            DimPlot_scCustom(df_result, group.by = "orig.ident") + NoAxes(),
            DimPlot(df_result, group.by = DF.name) + NoAxes())
  
  # Guardar el resultado en la lista
  df_results_list[[i]] <- df_result
}

```


```{r}
# Loop sobre la lista de resultados de dobletes de cada ratón
for (i in 1:length(df_results_list)) {
  # Obtener el resultado de los dobletes para el ratón actual
  df_result <- df_results_list[[i]]
  
  # Buscar el nombre del ratón
  raton_nombre <- unique(df_result@meta.data$orig.ident)
  
  # Buscar el nombre de la columna que contiene "DF.classifications"
  DF.name <- grep("DF.classifications", colnames(df_result@meta.data), value = TRUE)
  
  # Verificar si se encontró la columna
  if (length(DF.name) == 0) {
    cat("No se encontró la columna con el patrón 'DF.classifications' en el ratón", raton_nombre, "\n")
    next  # Pasar al siguiente ratón si no se encontró la columna
  }
  
  # Calcular el porcentaje de células clasificadas como dobletes
  num_doublets <- sum(df_result@meta.data[, DF.name] == "Doublet")
  total_cells <- nrow(df_result@meta.data)
  percent_doublets <- (num_doublets / total_cells) * 100
  
  # Imprimir resultados
  cat("En el ratón", raton_nombre, "el", percent_doublets, "% de las células fueron clasificadas como dobletes.\n")
  
  # Gráfico UMAP con etiquetas
  UMAPPlot(df_result, label = TRUE)
  
  # Gráfico de violín para comparar el número de genes detectados en dobletes y singletes
  VlnPlot(df_result, features = "nFeature_RNA", group.by = DF.name, pt.size = 0.1)
}
```

```{r}
# Para el primer ratón
raton_nombre_1 <- unique(df_results_list[[1]]@meta.data$orig.ident)
DF.name_1 <- grep("DF.classifications", colnames(df_results_list[[1]]@meta.data), value = TRUE)
plot_1_umap <- UMAPPlot(df_results_list[[1]], group.by = DF.name_1, label = TRUE)
plot_1_umap + ggtitle(paste("UMAP Plot - Ratón:", raton_nombre_1))
plot_1_vln <- VlnPlot(df_results_list[[1]], features = "nFeature_RNA", group.by = DF.name_1, pt.size = 0.1)
plot_1_vln + ggtitle(paste("Violin Plot - Ratón:", raton_nombre_1))

# Para el segundo ratón
raton_nombre_2 <- unique(df_results_list[[2]]@meta.data$orig.ident)
DF.name_2 <- grep("DF.classifications", colnames(df_results_list[[2]]@meta.data), value = TRUE)
plot_2_umap <- UMAPPlot(df_results_list[[2]], group.by = DF.name_2, label = TRUE)
plot_2_umap + ggtitle(paste("UMAP Plot - Ratón:", raton_nombre_2))
plot_2_vln <- VlnPlot(df_results_list[[2]], features = "nFeature_RNA", group.by = DF.name_2, pt.size = 0.1)
plot_2_vln + ggtitle(paste("Violin Plot - Ratón:", raton_nombre_2))

# Para el tercer ratón
raton_nombre_3 <- unique(df_results_list[[3]]@meta.data$orig.ident)
DF.name_3 <- grep("DF.classifications", colnames(df_results_list[[3]]@meta.data), value = TRUE)
plot_3_umap <- UMAPPlot(df_results_list[[3]], group.by = DF.name_3, label = TRUE)
plot_3_umap + ggtitle(paste("UMAP Plot - Ratón:", raton_nombre_3))
plot_3_vln <- VlnPlot(df_results_list[[3]], features = "nFeature_RNA", group.by = DF.name_3, pt.size = 0.1)
plot_3_vln + ggtitle(paste("Violin Plot - Ratón:", raton_nombre_3))

# Para el cuarto ratón
raton_nombre_4 <- unique(df_results_list[[4]]@meta.data$orig.ident)
DF.name_4 <- grep("DF.classifications", colnames(df_results_list[[4]]@meta.data), value = TRUE)
plot_4_umap <- UMAPPlot(df_results_list[[4]], group.by = DF.name_4, label = TRUE)
plot_4_umap + ggtitle(paste("UMAP Plot - Ratón:", raton_nombre_4))
plot_4_vln <- VlnPlot(df_results_list[[4]], features = "nFeature_RNA", group.by = DF.name_4, pt.size = 0.1)
plot_4_vln + ggtitle(paste("Violin Plot - Ratón:", raton_nombre_4))

# Para el quinto ratón
raton_nombre_5 <- unique(df_results_list[[5]]@meta.data$orig.ident)
DF.name_5 <- grep("DF.classifications", colnames(df_results_list[[5]]@meta.data), value = TRUE)
plot_5_umap <- UMAPPlot(df_results_list[[5]], group.by = DF.name_5, label = TRUE)
plot_5_umap + ggtitle(paste("UMAP Plot - Ratón:", raton_nombre_5))
plot_5_vln <- VlnPlot(df_results_list[[5]], features = "nFeature_RNA", group.by = DF.name_5, pt.size = 0.1)
plot_5_vln + ggtitle(paste("Violin Plot - Ratón:", raton_nombre_5))

# Para el sexto ratón
raton_nombre_6 <- unique(df_results_list[[6]]@meta.data$orig.ident)
DF.name_6 <- grep("DF.classifications", colnames(df_results_list[[6]]@meta.data), value = TRUE)
plot_6_umap <- UMAPPlot(df_results_list[[6]], group.by = DF.name_6, label = TRUE)
plot_6_umap + ggtitle(paste("UMAP Plot - Ratón:", raton_nombre_6))
plot_6_vln <- VlnPlot(df_results_list[[6]], features = "nFeature_RNA", group.by = DF.name_6, pt.size = 0.1)
plot_6_vln + ggtitle(paste("Violin Plot - Ratón:", raton_nombre_6))

```

# ELIMINAMOS LAS CÉLULAS PREDICHAS COMO DOBLETES E INTEGRAMOS

Si los eliminamos:

```{r}
# Crear una nueva lista para almacenar los resultados sin dobletes
df_results_list_singlets <- list()

# Iterar a través de cada ratón en la lista original
for (i in 1:length(df_results_list)) {
  # Obtener el nombre de la columna que contiene la clasificación de DoubletFinder
  DF.name <- grep("DF.classifications", colnames(df_results_list[[i]]@meta.data), value = TRUE)
  
  # Filtrar las filas donde la clasificación no es "Doublet"
  singlets_meta_data <- df_results_list[[i]]@meta.data[df_results_list[[i]]@meta.data[[DF.name]] != "Doublet", ]
  
  # Crear un nuevo objeto Seurat con las celdas filtradas
  singlets_object <- subset(df_results_list[[i]], cells = rownames(singlets_meta_data))
  
  # Añadir el objeto filtrado a la nueva lista
  df_results_list_singlets[[i]] <- singlets_object
}

# Mostrar el tamaño de cada objeto en la lista filtrada para verificación
for (i in 1:length(df_results_list_singlets)) {
  cat("Número de células en el ratón", unique(df_results_list_singlets[[i]]@meta.data$orig.ident), ":", nrow(df_results_list_singlets[[i]]@meta.data), "\n")
}
```

# INTEGRACION DE LOS DATOS CON RPCA SIN DOBLETES

```{r}
data_integrated_sin_dob <- integrate(df_results_list_singlets) 
```

```{r}
#data_integrated <- ScaleData(data_integrated,verbose = FALSE) #ya esta escalado, no estoy seguro pero si lo vuelvo a escalar se queda igual
#rm(mice.list)
DefaultAssay(data_integrated_sin_dob) <- "integrated"
data_integrated_sin_dob <- RunPCA(data_integrated_sin_dob, npcs = 20,verbose = FALSE)
data_integrated_sin_dob <- RunUMAP(data_integrated_sin_dob, dims = 1:20, reduction="pca")
data_integrated_sin_dob <- RunTSNE(data_integrated_sin_dob, dims = 1:20)
data_integrated_sin_dob <- FindNeighbors(data_integrated_sin_dob, dims = 1:20) %>% FindClusters(resolution = 0.2) #a high resolution parameter is used to get a finer clustering. 
```

```{r}
VizDimLoadings(data_integrated_sin_dob, dims = 1:10, nfeatures = 20, reduction = "pca", ncol=4)
```

```{r}
ElbowPlot(data_integrated_sin_dob, ndims = 20) + geom_hline(yintercept = 2, linetype = "dashed", color = "red") + geom_hline(yintercept = 3, linetype = "dashed", color = "blue")
```

```{r}
DimPlot(data_integrated_solo_BCR, split.by = "orig.ident")   
```

```{r}
DimPlot_scCustom(data_integrated_sin_dob,reduction = "umap", group.by = "orig.ident",
                 colors_use = DiscretePalette_scCustomize(num_colors = 6, palette = "polychrome", shuffle_pal = TRUE))
DimPlot_scCustom(data_integrated_sin_dob,reduction = "pca",group.by = "orig.ident",
                 colors_use = DiscretePalette_scCustomize(num_colors = 6, palette = "polychrome", shuffle_pal = TRUE))
DimPlot_scCustom(data_integrated_sin_dob,reduction = "tsne", group.by = "orig.ident",
                 colors_use = DiscretePalette_scCustomize(num_colors = 6, palette = "polychrome", shuffle_pal = TRUE))
UMAPPlot(data_integrated_sin_dob, label= T)
TSNEPlot(data_integrated_sin_dob, label=T)
```

```{r}
PCA_cc <- RunPCA(data_integrated_sin_dob, features = c(m.s.genes, m.g2m.genes))

DimPlot_scCustom(PCA_cc,reduction = "pca",group.by = "Phase",
                 colors_use = DiscretePalette_scCustomize(num_colors = 6, palette = "polychrome", shuffle_pal = TRUE))

DimPlot_scCustom(data_integrated_sin_dob,reduction = "umap",group.by = "Phase",
                 colors_use = DiscretePalette_scCustomize(num_colors = 6, palette = "polychrome", shuffle_pal = TRUE))

DimPlot_scCustom(data_integrated_sin_dob,reduction = "tsne",group.by = "Phase",
                 colors_use = DiscretePalette_scCustomize(num_colors = 6, palette = "polychrome", shuffle_pal = TRUE))

DimPlot_scCustom(data_integrated_sin_dob,reduction = "umap", group.by = "Phase", split.by = "seurat_clusters",
                 colors_use = DiscretePalette_scCustomize(num_colors = 6, palette = "polychrome", shuffle_pal =TRUE))
```

# Células de cada cluster sin DOBLETES:

```{r}
ob.list <- SplitObject(data_integrated_sin_dob, split.by = "orig.ident")
plot.list <- lapply(X = ob.list, FUN = function(x) {
    DimPlot(x, reduction = "umap", label = F,label.size = 4.75,pt.size=1.1,label.color = "black")})
plot.list
```

```{r}
ob.list <- SplitObject(data_integrated_sin_dob, split.by = "orig.ident")
plot.list <- lapply(X = ob.list, FUN = function(x) {
    DimPlot(x, reduction = "tsne", label = F,label.size = 4.75,pt.size=1.1,label.color = "black")})
plot.list
```


# Guardamos el archivo para analiza el repertorio

```{r}
saveRDS(data_integrated_sin_dob, file="scRNA_con_BCR_para_repertorio.nf.sin-dobletes.rds")
```
