==> Analysis_sc_R/04_Analisis_firmas_transcriptómicas_nf.Rmd <==
---
title: "Analisis_firmas_transcriptomico_nf"
author: "Enric Vercher"
date: "2024-06-03"
output: html_document
---

```{r}
suppressPackageStartupMessages({
  library(BiocParallel)
  library(clusterProfiler)
  library(ComplexHeatmap)
  library(DOSE)
  library(DoubletFinder)
  library(DropletUtils)
  library(dplyr)
  library(enrichplot)
  library(fgsea)
  library(glmGamPoi)
  library(ggplot2)
  library(heatmap3)
  library(knitr)
  library(magrittr)
  library(msigdbr)
  library(Nebulosa)
  library(org.Mm.eg.db)
  library(parallel)
  library(patchwork)
  library(presto)
  library(qs)
  library(ReactomePA)
  library(scCustomize)
  library(scRepertoire)
  library(scales)
  library(sctransform)
  library(Seurat)
  library(SeuratData)
  library(SeuratDisk)
  library(SeuratWrappers)
  library(SignatuR)
  library(singleseqgset)
  library(SoupX)
  library(splatter)
  library(stats)
  library(tidyverse)
  library(UCell)
  library(viridis)
})
```

Cargamos el archivo con los 5200 células con el BCR

```{r}
data.RNA.BCR.filt.wo.dob <- readRDS(file="C:/Users/d940401/Desktop/Analisis_Seurat/scRNA_con_BCR_para_repertorio.nf.sin-dobletes.rds")
```


```{r}
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
plot1 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, group.by="orig.ident")
plot2 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, label = T)
plot3 <-FeaturePlot_scCustom(data.RNA.BCR.filt.wo.dob, features = c("Cd79a", "Sdc1", "Cd19", "Xbp1"), pt.size = 0.1,order = F, na_cutoff = NA,alpha_exp = 0.75)
((plot1 / plot2) | plot3) + plot_layout(width = c(1,2))
```

```{r}
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
plot1 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, group.by="orig.ident")
plot2 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, label = T)
plot3 <-FeaturePlot_scCustom(data.RNA.BCR.filt.wo.dob, features = c("Cd19","Ptprc","Ms4a1","Cd38","Cd80","Sdc1"), pt.size = 0.1,order = F, na_cutoff = NA,alpha_exp = 0.75)
((plot1 / plot2) | plot3) + plot_layout(width = c(1,2))
```

https://github.com/satijalab/seurat/issues/4082
https://github.com/satijalab/seurat/issues/1717

To keep this simple: You should use the integrated assay when trying to 'align' cell states that are shared across datasets (i.e. for clustering, visualization, learning pseudotime, etc.) You should use the RNA assay when exploring the genes that change either across clusters, trajectories, or conditions.

We don't suggest you use the integrated data slot for the visualization. The integrated data is mainly aimed for cell type identification.

```{r}
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
FeaturePlot(data.RNA.BCR.filt.wo.dob, c("Cd79a","Xbp1","Cd80","Sell","Foxo1","Cd93","Cd19","C1qa","Cd1d1"),
                     ncol=3, reduction = "umap",min.cutoff = 0)
```

```{r}
plot1 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, group.by="orig.ident")
plot2 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, label = T)
plot3 <- FeaturePlot_scCustom(data.RNA.BCR.filt.wo.dob, features = c("Cd79a","Sdc1","Bach2","Sell","Foxo1","Cd93","Cd19","C1qa","C1qb"), pt.size = 0.1,order = F, na_cutoff = NA,alpha_exp = 0.75)
((plot1 / plot2) | plot3) + plot_layout(width = c(1,2))
```

Marcadores asociados a poblaciones mieloides:

```{r}
FeaturePlot(data.RNA.BCR.filt.wo.dob, c("Ctsb","Cd68","Selenop","Msr1","Spp1","Csf1r","Cxcl2","Il1b","Pltp","Cxcl10","Cd14","Cst3"),
                     ncol=3, reduction = "umap")
```

```{r}
plot1 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, group.by="orig.ident")
plot2 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, label = T)
plot3 <- FeaturePlot_scCustom(data.RNA.BCR.filt.wo.dob, features = c("Ctsb","Cd68","Selenop","Msr1","Csf1r","Cxcl2","Il1b","Cxcl10","Cd14","Cst3"), pt.size = 0.1,order = F, na_cutoff = NA,alpha_exp = 0.75)
((plot1 / plot2) | plot3) + plot_layout(width = c(0.5,1))
```

```{r}
FeaturePlot(data.RNA.BCR.filt.wo.dob, c("Lst1","Aif1","Csf1r","Fcgr3","Cd1d1","Klrd1"),
                     ncol=3, reduction = "umap")
```


```{r}
plot1 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, group.by="orig.ident")
plot2 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, label = T)
plot3 <- FeaturePlot_scCustom(data.RNA.BCR.filt.wo.dob, features =  c("Lst1","Aif1","Csf1r","Fcgr3","Cd1d1","Klrd1"), pt.size = 0.1,order = F, na_cutoff = NA,alpha_exp = 0.75)
((plot1 / plot2) | plot3) + plot_layout(width = c(0.5,1))
```


```{r}
FeaturePlot(data.RNA.BCR.filt.wo.dob, c("Cd79a","Cd79b","Ighm","Ighd","Cd69","Cd83","Sdc1"),
                     ncol=3, reduction = "umap")
```

CD79a se expresa en casi todas las etapas de desarrollo de células B, desde pro-B hasta células B maduras, y disminuye en las células B secretoras.(https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8555128/)

Se acompaña de una disminuación de CD19, 

```{r}
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
VlnPlot(data.RNA.BCR.filt.wo.dob, features= c("Cd79a","Cd79b","Cd69","Cd83","Sdc1","Cd19"),
                     ncol=3) 
```

```{r}
FeaturePlot(data.RNA.BCR.filt.wo.dob, c("Junb","Fcer2a","Sell","Ptprc","H2-DMb2","H2-Eb2","H2-Ab1","H2-Ob","H2-Aa"),
                     ncol=3, reduction = "umap")
```

Ptprc --> memoria y naive.

Activados:

```{r}
VlnPlot(data.RNA.BCR.filt.wo.dob,features=c("Junb","Fcer2a","Sell","Ptprc","H2-DMb2","H2-Eb2","H2-Ab1","H2-Ob","H2-Aa")) 
```


```{r}
FeaturePlot(data.RNA.BCR.filt.wo.dob, c("Ebf1", "Pax5", "Foxo1","Bach2","Cd38","Zbtb20","Ptprc","Cd86"),
                     ncol=3, reduction = "umap")
```

```{r}
plot1 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, group.by="orig.ident")
plot2 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, label = T)
plot3 <- FeaturePlot_scCustom(data.RNA.BCR.filt.wo.dob, features =  c("Ebf1", "Pax5", "Foxo1","Bach2","Cd38","Zbtb20","Ptprc","Cd86"), pt.size = 0.1,order = F, na_cutoff = NA,alpha_exp = 0.75)
((plot1 / plot2) | plot3) + plot_layout(width = c(0.5,1))
```

```{r}
VlnPlot(data.RNA.BCR.filt.wo.dob,features=c("Ebf1", "Pax5", "Foxo1","Bach2","Cd38","Zbtb20","Ptprc","Cd86")) 
```

```{r}
VlnPlot(data.RNA.BCR.filt.wo.dob,features=c("Glrx","Blk","Cd80","Cd274")) 
```

Clúster 0,1,2 son naive y activadas --> naive-activadas, activadas, activadas-prememoria, prememoria, memoria 1 y memoria 2.
El clúster 3: prememoria.
Clúster 4 plasmablastos? por el GLrx. Menos CD19 y menos CD79a, lo ha perido pero aun no es 0.
Blk --> en memoria mas que naive y mas switched que no. 
8 y 9 diria que son exhausted switched B cell. Cd80 muy alto. Cd274 alto (PDL1) o plasmablastos más completos y ya el 7 plasmáticas.


Memoria:

```{r}
FeaturePlot(data.RNA.BCR.filt.wo.dob,c("Cd80", "Cd84", "Cd86","Pdcd1lg2","Itga4","Traf1","Cd38"),
                     ncol=3, reduction = "umap")
```

```{r}
plot1 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, group.by="orig.ident")
plot2 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, label = T)
plot3 <- FeaturePlot_scCustom(data.RNA.BCR.filt.wo.dob, features =  c("Cd80", "Cd84", "Cd86","Pdcd1lg2","Itga4","Traf1","Cd38"), pt.size = 0.1,order = F, na_cutoff = NA,alpha_exp = 0.75)
((plot1 / plot2) | plot3) + plot_layout(width = c(0.5,1))
```

```{r}
FeaturePlot(data.RNA.BCR.filt.wo.dob,c("Sdc1","Prdm1","Irf4","Xbp1"),
                     ncol=3, reduction = "umap")
```

```{r}
plot1 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, group.by="orig.ident")
plot2 <- UMAPPlot(data.RNA.BCR.filt.wo.dob, label = T)
plot3 <- FeaturePlot_scCustom(data.RNA.BCR.filt.wo.dob, features = c("Sdc1","Prdm1","Irf4","Xbp1"), pt.size = 0.1,order = F, na_cutoff = NA,alpha_exp = 0.75)
((plot1 / plot2) | plot3) + plot_layout(width = c(0.5,1))
```

```{r}
FeaturePlot_scCustom(data.RNA.BCR.filt.wo.dob, features =  c("Ms4a1","Jchain","Cd83","Cd69", "Vpreb3"), pt.size = 0.1,order = F, na_cutoff = NA,alpha_exp = 0.75)
```
```{r}
FeaturePlot_scCustom(data.RNA.BCR.filt.wo.dob, features =  c("Cd19","Cd80","Pdcd1lg2","Ptprc","Cd38","Nt5e","Sdc1"), pt.size = 0.1,order = F, na_cutoff = NA,alpha_exp = 0.75)
```

```{r}
VlnPlot(data.RNA.BCR.filt.wo.dob,features=c("Cd19","Cd80","Pdcd1lg2","Ptprc","Cd38","Nt5e","Sdc1")) 

```

# Firmas transcriptómicas

```{r}
markers <- list()
markers$naive <- c("Cd79a", "Cd79b","Cd19")
#markers$naive2 <- c("Gpr17", "Kcng1", "Dsp", "Apbb2", "Jsrp1", "Slc26a4", "Tubb3", "Ccn2", "Akap6", "Slc2a5", "H4c4", "Dbndd1"). No sirven
#markers$memory2 <- c("Tas1r3", "Arl14", "Apod", "Rassf6", "Il5", "Dkkl1", "H3c7", "Spaca3", "Slc16a4", "Ntng1", "Pcdhga12")
markers$activated <- c("Junb","Fcer2a","Sell","Ptprc","H2-DMb2","H2-Eb2","H2-Ab1","H2-Ob","H2-Aa")
markers$prememory <- c("Ebf1", "Pax5", "Foxo1","Bach2","Cd38","Zbtb20","Ptprc","Cd86")
markers$memory <- c("Cd80", "Cd84", "Cd86","Pdcd1lg2","Itga4","Traf1","Cd38")
markers$plasmablast <- c("Bcl6","Havcr2","Cd93","Pdcd1lg2","Glrx")
markers$plasmatic <- c("Sdc1","Prdm1","Irf4","Xbp1")
#markers$mieloides <- c("Ctsb","Cd68","Selenop","Msr1","Spp1","Csf1r","Cxcl2","Il1b","Pltp","Cxcl10","Cd14","Cst3")
```

```{r}
data.RNA.BCR.filt.wo.dob <- JoinLayers(data.RNA.BCR.filt.wo.dob)
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
data.RNA.BCR.filt.wo.dob<- AddModuleScore_UCell(data.RNA.BCR.filt.wo.dob, features = markers)
signature.names <- paste0(names(markers), "_UCell")
VlnPlot_scCustom(data.RNA.BCR.filt.wo.dob,plot_boxplot=TRUE,features = signature.names, group.by ="seurat_clusters")
```

Más smooth:

```{r}
smoth_seurat <- SmoothKNN(data.RNA.BCR.filt.wo.dob,
                           signature.names = signature.names,
                           reduction="pca")
FeaturePlot(smoth_seurat, reduction = "umap", features = signature.names)
```

```{r}
FeaturePlot(data.RNA.BCR.filt.wo.dob, reduction = "umap", features = signature.names)
```

## Mas firmas transcriptómicas


```{r}
markers <- list()
markers$naive <- c("Cd79a", "Cd79b","Cd19")
#markers$naive2 <- c("Gpr17", "Kcng1", "Dsp", "Apbb2", "Jsrp1", "Slc26a4", "Tubb3", "Ccn2", "Akap6", "Slc2a5", "H4c4", "Dbndd1"). No sirven
#markers$memory2 <- c("Tas1r3", "Arl14", "Apod", "Rassf6", "Il5", "Dkkl1", "H3c7", "Spaca3", "Slc16a4", "Ntng1", "Pcdhga12")
markers$activated <- c("Junb","Fcer2a","Sell","Ptprc","H2-DMb2","H2-Eb2","H2-Ab1","H2-Ob","H2-Aa")
markers$memory <- c("Cd80", "Cd84", "Cd86","Pdcd1lg2","Itga4","Traf1","Cd38","Ebf1","Pax5", "Foxo1","Bach2","Cd38","Zbtb20","Ptprc","Cd86")
markers$plasmablast <- c("Bcl6","Havcr2","Cd93","Pdcd1lg2","Glrx")
markers$plasmatic <- c("Sdc1","Prdm1","Irf4","Xbp1")
#markers$mieloides <- c("Ctsb","Cd68","Selenop","Msr1","Spp1","Csf1r","Cxcl2","Il1b","Pltp","Cxcl10","Cd14","Cst3")

data.RNA.BCR.filt.wo.dob <- JoinLayers(data.RNA.BCR.filt.wo.dob)
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
data.RNA.BCR.filt.wo.dob<- AddModuleScore_UCell(data.RNA.BCR.filt.wo.dob, features = markers)
signature.names <- paste0(names(markers), "_UCell")
VlnPlot_scCustom(data.RNA.BCR.filt.wo.dob,plot_boxplot=TRUE,features = signature.names, group.by ="seurat_clusters")

FeaturePlot(data.RNA.BCR.filt.wo.dob, reduction = "umap", features = signature.names)
```

# para las naive:

https://github.com/satijalab/seurat/issues/3521

```{r}
# Define the list of marker genes for Naive
DefaultAssay(data_integrated_VDJ) <- "RNA"
naive_marker_gene_list <- list(c("Ighd", "Cd79a", "Cd79b"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data_integrated_VDJ, features = naive_marker_gene_list, name = "Naive")

FeaturePlot(object = object, reduction="umap",features = "Naive1", pt.size = 0.9) +  scale_colour_gradientn(colors = c("white", "blue"), na.value = "grey90", limits = c(-4, 10))
```

```{r}
# Define the list of marker genes for activated
DefaultAssay(data_integrated_VDJ) <- "RNA"
activated_marker_gene_list <- list(c("Junb","Fcer2a","Sell","Ptprc","H2-DMb2","H2-Eb2","H2-Ab1","H2-Ob","H2-Aa"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data_integrated_VDJ, features = activated_marker_gene_list, name = "activated")

FeaturePlot(object = object, features = "activated1", pt.size = 0.9) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(-4, 10))
```

```{r}
# Define the list of marker genes for prememory
DefaultAssay(data_integrated_VDJ) <- "RNA"
prememory_marker_gene_list <- list(c("Ebf1", "Pax5", "Foxo1","Bach2","Cd38","Zbtb20","Ptprc","Cd86"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data_integrated_VDJ, features = prememory_marker_gene_list, name = "Prememory")

FeaturePlot(object = object, features = "Prememory1", pt.size = 0.9) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(-4, 10))
```

#Cluster de memoria:

```{r}
# Define the list of marker genes for Naive
DefaultAssay(data_integrated_VDJ) <- "RNA"
memory_marker_gene_list <- list(c("Cd80", "Cd84", "Cd86","Pdcd1lg2","Itga4","Traf1","Ighg3","Cd38"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data_integrated_VDJ, features = memory_marker_gene_list, name = "memory")

FeaturePlot(object = object, features = "memory1", pt.size = 0.9) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(-4, 10))
```

```{r}
# Define the list of marker genes for plasmatic
DefaultAssay(data_integrated_VDJ) <- "RNA"
plasmatic_marker_gene_list <- list(c("Sdc1","Prdm1","Irf4","Xbp1"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data_integrated_VDJ, features = plasmatic_marker_gene_list, name = "plasmic")

FeaturePlot(object = object, features = "plasmic1", pt.size = 1) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(-1, 10))
```

```{r}
# plasmablasto
#regulador
# Define the list of marker genes for plasmablast
DefaultAssay(data_integrated_VDJ) <- "RNA"
plasmablast_marker_gene_list <- list(c("Bcl6","Havcr2","Cd93","Pdcd1lg2"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data_integrated_VDJ, features = plasmablast_marker_gene_list, name = "plasmablasto")

FeaturePlot(object = object, features = "plasmablasto1", pt.size = 1) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(-1, 10))

```

```{r}
#regulador
# Define the list of marker genes for plasmatic
DefaultAssay(data_integrated_VDJ) <- "RNA"
regulator_marker_gene_list <- list(c("Il12a","Cd24a"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data_integrated_VDJ, features = regulator_marker_gene_list, name = "regulator")

FeaturePlot(object = object, features = "regulator1", pt.size = 1) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(-1, 10))
```

# Analisis DEG

```{r}
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
data_integrated_sin_dob <- JoinLayers(data.RNA.BCR.filt.wo.dob)
#data.RNA.BCR.filt.wo.dob <- PrepSCTFindMarkers(data.RNA.BCR.filt.wo.dob)
m.all.markers <- FindAllMarkers(data_integrated_sin_dob, only.pos = TRUE, min.pct = 0.25, logfc.threshold = 0.25)
m.all.markers %>%
    group_by(cluster) %>%
    slice_max(n = 10, order_by = avg_log2FC)
#write.table(m.all.markers, "C:/Users/d940401/Desktop/Analisis_Seurat/all-markers-DE-CCA-sin-Cd8.tsv", sep = "/t", row.names = FALSE)

```

```{r}
top10_cl_markers <- m.all.markers %>% group_by(cluster) %>% top_n(n = 10, wt = avg_log2FC)
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
DoHeatmap(data.RNA.BCR.filt.wo.dob, features = top10_cl_markers$gene,assay="integrated") + NoLegend()
```

```{r}
DotPlot_scCustom(data.RNA.BCR.filt.wo.dob, features = c("Cd79a", "Cd79b","Ebf1", "Pax5", "Foxo1","Bach2","Sdc1","Prdm1","Irf4","Xbp1","Cd68","Csf1r","C1qa"),flip_axes=TRUE) 
```

```{r}
# Find markers and limit to those expressed in greater than 75% of target population
# https://samuel-marsh.github.io/scCustomize/articles/Gene_Expression_Plotting.html
all_markers <- FindAllMarkers(object = data.RNA.BCR.filt.wo.dob) %>%
    Add_Pct_Diff() %>%
    filter(pct_diff > 0.6)
top_markers <- Extract_Top_Markers(marker_dataframe = all_markers, num_genes = 7, named_vector = FALSE,
    make_unique = TRUE)

Clustered_DotPlot(seurat_object = data.RNA.BCR.filt.wo.dob, features = top_markers,k = 7)
```

Algo extra:

```{r}
# Find markers and limit to those expressed in greater than 75% of target population
# https://samuel-marsh.github.io/scCustomize/articles/Gene_Expression_Plotting.html
all_markers <- FindAllMarkers(object = all_mice_RPCA_integrated_sin_CD8_doublet) %>%
    Add_Pct_Diff() %>%
    filter(pct_diff > 0.6)
top_markers <- Extract_Top_Markers(marker_dataframe = all_markers, num_genes = 7, named_vector = FALSE,
    make_unique = TRUE)

Clustered_DotPlot(seurat_object = all_mice_RPCA_integrated_sin_CD8_doublet, features = top_markers,k = 7)
```

# Apellidos a los clústeres

```{r}
head(Idents(data.RNA.BCR.filt.wo.dob), 5)
new.cluster.ids <-  c("C0-Naive/Naive activada", "C1-Plasmblasto","C2-memoria (no-IS)","C3-act +PreMem","C4-Memoria+plasmblasto","C5-Bmemoria-IS","C6-Bmemoria-IS","C7-Naive","C8-Plasmblasto-avanzado","C9-Plasmatica","C10-Breg(IS)")
names(new.cluster.ids) <- levels(data.RNA.BCR.filt.wo.dob)
data.RNA.BCR.filt.wo.dob <- RenameIdents(data.RNA.BCR.filt.wo.dob, new.cluster.ids)
head(Idents(data.RNA.BCR.filt.wo.dob), 5)
```


```{r}
new.cluster.ids <- c("C0", "C1","C2","C3","C4","C5","C6","C7","C8","C9","C10")
legend.ids <- c("C0-Naive/Naive activada", "C1-Plasmblasto","C2-memoria (no-IS)","C3-act +PreMem","C4-Memoria+plasmblasto","C5-Bmemoria-IS","C6-Bmemoria-IS","C7-Naive","C8-Plasmblasto-avanzado","C9-Plasmatica","C10-Breg(IS)")
names(new.cluster.ids) <- levels(data.RNA.BCR.filt.wo.dob)
data.RNA.BCR.filt.wo.dob <- RenameIdents(data.RNA.BCR.filt.wo.dob, new.cluster.ids)
DimPlot(data.RNA.BCR.filt.wo.dob, reduction = "umap", label = TRUE, pt.size = 1.5) 
```

# Las naive/naive-activadas y premem

```{r}
# Set color palette
pal <- viridis(n = 2, option = "C", direction = -1)
B_activated <-c("Junb","Fcer2a","Sell")
b_cell_markers <- c("Ebf1", "Pax5", "Foxo1","Bach2","Plac8","Ighm","Bcl6","Vim","Cd79a","Ighd","Cd69","Ccr7")
#FeaturePlot(data.RNA.BCR.filt.wo.dob, features = b_cell_markers,cols = pal, order = T)
VlnPlot(object = data.RNA.BCR.filt.wo.dob, features = b_cell_markers,pt.size=0, idents=c("0","5","9"))+ geom_boxplot(width=0.1, color="black", alpha=0.2)
DotPlot(data.RNA.BCR.filt.wo.dob, features = c(b_cell_markers,B_activated), group.by = "seurat_clusters") + coord_flip() 
```

Comparación con los datos de citometría

esto con lo de Sandra del power point y con esta web:

https://www.novusbio.com/research-areas/immunology/b-cell-development-and-differentiation-markers

```{r}
# B220 o CD45R es Ptprc
# CD138 es Sdc1
# CD20 es Ms4a1
# Havcr2 es TIM3
# Ms4a1 es CD20
# Il12a es IL35
# Pdcd1lg2 es CD273
# Nt5e es CD73
# il6ra es CD126
# cxcr4 es CD184
# slamf7 es CD319
b_cell_cytometry <- c("Ptprc","Cd38","Sdc1","Xbp1","Jchain","Ebf1", "Pax5", "Foxo1","Bach2","Cd80", 
                      "Cd86","Ighg2c","Ighg2b","Ighg3","Ighg1","Ighm","Ighd","Igha","Cd1d1","Havcr2","Il12a","Cd24a",
                      "Ms4a1","H2-DMb2","H2-Eb2","H2-Ab1","H2-Ob","H2-Aa","Cd93","Apoe","Plac8","Vim",
                      "Cd84","Pdcd1lg2","Nt5e","Il6ra","Cxcr4","Slamf7","Cd9","Traf1","Itga4","Zbtb20","Ccr6")
DotPlot(data.RNA.BCR.filt.wo.dob, features = b_cell_cytometry, group.by = "seurat_clusters") + coord_flip() 
```


https://www.nature.com/articles/s41581-018-0074-7/figures/1

con el poster: https://www.bio-rad-antibodies.com/static/2016/mouse-human/mouse-immune-cell-lineage--antigen-expression-vh.pdf

# Marcadores para el TFM

```{r}
b_cell_cytometry_figura <- c("Cd24a", "Il12a", "Igha", "Ighg2c", "Ighg2b", "Ighg1", "Bcl6","Sdc1","Jchain","Xbp1","Nt5e", "Pdcd1lg2", "Cd86", "Traf1", "Itga4", "Cd84", "Cd80", "Cd38", "Zbtb20", "Pax5", "Bach2", "Ebf1", "Foxo1", "Cd1d1", "Ptprc", "H2-Aa", "H2-Ob", "H2-Ab1", "H2-Eb2", "H2-DMb2", "Havcr2", "Cd93","Apoe", "Junb", "Fcer2a", "Sell", "Cd79b", "Cd79a", "Ighd", "Ighm")
DotPlot(data.RNA.BCR.filt.wo.dob, features = b_cell_cytometry_figura, group.by = "seurat_clusters") + coord_flip() 
```

Naive: Cluster 0 hay polbacion naive (más naive)


B activadas: cluster 3 --> Ebf1 y Pax5, Foxo1 y Bmemoria (prememoria por Foxo1 y Ebf1, Bach2). En este caso el cluster 2 parece ser más memoria que el 3. Se expresa menos el Foxo1 y el Ebf1, por lo que son más Bmem que el cluster 3. el 5 son más memoria porque estos marcadores desaparecen. Hay una población naive también. El cluster 7 son naive 

del bhattacharya: A recent study found more heterogeneous expression of CD80 in memory B cells than we observed in this study, with the CD80 cells representing cells that had undergone minimal somatic hypermutation. 

```{r}
VlnPlot(object = data.RNA.BCR.filt.wo.dob, features =c("Cd1d1","Il12a","Ptprc","Cd80","Cd24a","","Havcr2","Pdcd1lg2","Ms4a1"),pt.size=0)
```


```{r}
FeaturePlot_scCustom(data.RNA.BCR.filt.wo.dob, features = c("Cd79a", "Sdc1", "Ptprc", "Cd80","H2-Eb2","Havcr2"), pt.size = 0.1)
VlnPlot(data.RNA.BCR.filt.wo.dob, features = c("Pdcd1lg2", "Cd93", "Sdc1","Il12a"),pt.size=0)
```

```{r}
VlnPlot(data.RNA.BCR.filt.wo.dob, features = c("Srm","Ncl"),pt.size=0)
#naive
```

https://www.nature.com/articles/ni.2641

El Junb está en las naive y las naive activadas.

 migration/adhesion

```{r}
memory <- c("Ski","Mll3","PmlTcf4","Bmi1")
DotPlot(data.RNA.BCR.filt.wo.dob, features = memory, group.by = "seurat_clusters") + coord_flip() 
```


# Esto del power point de Sandra
 
Si queremos ver si son más o menos memoria.

El cluster 9 parece el más plasmático al expresar el Sdc1 (CD138), Xbp1, Jchain, Slpi. Los mismo que en el paper: https://doi.org/10.1016/j.celrep.2021.109286


```{r}
#MemoryCells del paper 
Bmem <- c("Apoe","Phf21a", "Zbtb20", "Kmt2c", "Ski", "Tcf4", "Cd80", "Il6ra", "Itgb7", "Ccr6", "Pecam1", "Nid1","Ackr3", "Itga4", "Bcl2", "Traf1", "Malt1", "Ccnd3", "Rb1")
DotPlot(data.RNA.BCR.filt.wo.dob, features = Bmem, group.by = "seurat_clusters") + coord_flip()
```


```{r}
DotPlot(data.RNA.BCR.filt.wo.dob, features = PC, group.by = "seurat_clusters") + coord_flip() 
DotPlot(data.RNA.BCR.filt.wo.dob, features = Bmem, group.by = "seurat_clusters") + coord_flip() 
```

```{r}
# No había cd27
# No había cd73
# No había PDL1
plasmablast <- c("Ptprc","Cd24a","Cd38","Sdc1","Cd80")
DotPlot(data.RNA.BCR.filt.wo.dob, features = plasmablast, group.by = "seurat_clusters") + coord_flip() 
```
```{r}
#GerminalCenter can be subdivided into Light zone (LZ) and Dark Zone (DZ)

DZ.genes <- c("Pif1", "Otub2", "Psrc1", "Lrrc49", "Ccnb2", "Mnd1", "Lgr5", "Adhfe1", "Pafah1b3", "Hmmr", "Tifa", "Reln", "Sepp1", "Cdkn3", "Gcnt3", "Cdc20", "Ube2c", "Serinc5", "Cenpa", "Ccnb1", "Plk1", "Cenpf", "Tgfa", "Tpx2", "Depdc1b", "Gpd1l", "Polh", "Cep55", "Nek2", "Zkscan16", "Lmo4", "Gas2l3", "Ptgr1", "Aspm", "Rnf125", "Cxcr4", "Kif20a", "Kif2c", "Rabgap1l", "Bub1", "Nebl", "Lig4", "Neil1", "Pde2a", "Ehf", "Racgap1", "Cdca8", "Cdc14b", "Cdc25c", "Lipc", "Mprip", "Aurka", "Gpsm2", "Spdya", "Akap12", "Kif22", "Birc5", "Pfn2", "Ddx19a", "Ldlrad4", "Gcsam")

DotPlot(data.RNA.BCR.filt.wo.dob, features = DZ.genes, group.by = "seurat_clusters") + coord_flip()
```


```{r}
LZ.genes <- c("Egr1", "Egr2", "Ccnd2", "Cd69", "Cd86", "Cd40", "Slamf1", "Ifi30", "Myc","S1pr3", "Il4i1", "Ccr6", "", "Gpr183", "Nfkbia", "Plscr1", "Socs3", "Lifr", "Bcl2a1a", "Bcl2a1d", "Bcl2a1b", "Hhex", "Dock10", "Hivep3", "Samsn1", "Irf4", "Gimap4", "Ptger4",  "Fscn1", "Marcks", "Ier2", "Snn", "Bhlhe40", "Slpi", "Matr3", "Rapgef4", "Rilpl2", "Gadd45g", "Cfp", "Tyms", "Mdn1")

# DZ and LZ can be further roughly subdivided depending on cell cycle score and function
# Generally light zone cells are in G1 and S. DZ cellsare in S, G2/M and G1

# LZ cells in S are entering dark zone
# DZ cells in G1 are exiting DZ and entering LZ

#LZcells receiving Tfh cells signals to enter/exit the GC
LZ.help <- c("Itgb2", "Cxcr5", "Tgfbr2", "Il21r", "Ebf1", "Irf8", "Nfkb2", "Icosl", "Slamf1", "Tnfrsf13c")

#Cell in the lightzone Reentering to DZ
DZ.reentry <- c("Aicda", "Cxcr4", "Xrcc4", "Cd24a", "Mki67", "Foxo1", "Tcf3")

#Pre-memory cells
Premem <- c("Pbx3", "Ncoa1", "Etv6", "Efnb1", "S1pr3", "Bmpr1a", "Cd86", "Cd59a")
```


# Analisis FIRMAS TRANSCRIPCIONALES 2

# para las naive:

https://github.com/satijalab/seurat/issues/3521

```{r}
# Define the list of marker genes for Naive
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
naive_marker_gene_list <- list(c("Cd79a", "Cd79b","Cd19"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data.RNA.BCR.filt.wo.dob, features = naive_marker_gene_list, name = "Naive")

FeaturePlot(object = object, reduction="umap",features = "Naive1", pt.size = 0.9) +  scale_colour_gradientn(colors = c("white", "blue"), na.value = "grey90", limits = c(0, 4))
```

```{r}
# Define the list of marker genes for activated
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
activated_marker_gene_list <- list(c("Junb","Fcer2a","Sell","Ptprc","H2-DMb2","H2-Eb2","H2-Ab1","H2-Ob","H2-Aa"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data.RNA.BCR.filt.wo.dob, features = activated_marker_gene_list, name = "activated")

FeaturePlot(object = object, features = "activated1", pt.size = 0.9) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(0, 4))
```

```{r}
# Define the list of marker genes for prememory
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
prememory_marker_gene_list <- list(c("Ebf1", "Pax5", "Foxo1","Bach2","Cd38","Zbtb20","Ptprc","Cd86"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data.RNA.BCR.filt.wo.dob, features = prememory_marker_gene_list, name = "Prememory")

FeaturePlot(object = object, features = "Prememory1", pt.size = 0.9) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(0, 4))
```

#Cluster de memoria:

```{r}
# Define the list of marker genes for Naive
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
memory_marker_gene_list <- list(c("Cd80", "Cd84", "Cd86","Pdcd1lg2","Itga4","Traf1","Cd38"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data.RNA.BCR.filt.wo.dob, features = memory_marker_gene_list, name = "memory")

FeaturePlot(object = object, features = "memory1", pt.size = 0.9) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(0, 3))
```

```{r}
# Define the list of marker genes for plasmatic
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
plasmatic_marker_gene_list <- list(c("Sdc1","Prdm1","Irf4","Xbp1"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data.RNA.BCR.filt.wo.dob, features = plasmatic_marker_gene_list, name = "plasmic")

FeaturePlot(object = object, features = "plasmic1", pt.size = 1) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(0, 3))
```

```{r}
# plasmablasto
#regulador
# Define the list of marker genes for plasmablast
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
plasmablast_marker_gene_list <- list(c("Havcr2","Cd93","Pdcd1lg2","Glrx"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data.RNA.BCR.filt.wo.dob, features = plasmablast_marker_gene_list, name = "plasmablasto")

FeaturePlot(object = object, features = "plasmablasto1", pt.size = 1) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(0,2))

```

```{r}
#regulador
# Define the list of marker genes for plasmatic
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
regulator_marker_gene_list <- list(c("Il12a","Cd24a"))

# Add the module score to the Seurat object
object <- AddModuleScore(object = data.RNA.BCR.filt.wo.dob, features = regulator_marker_gene_list, name = "regulator")

FeaturePlot(object = object, features = "regulator1", pt.size = 1) +  scale_colour_gradientn(colors = c("white", "blue"),na.value = "grey90", limits = c(-1, 10))
```

```{r}
markers <- list()
markers$naive <- c("Ighd+", "Cd79a+", "Cd79b+")
markers$activated <- c("Junb","Fcer2a","Sell","Ptprc","H2-DMb2","H2-Eb2","H2-Ab1","H2-Ob","H2-Aa")
markers$prememory <- c("Ebf1", "Pax5", "Foxo1","Bach2","Cd38","Zbtb20","Ptprc","Cd86")
markers$memory <- c("Cd80", "Cd84", "Cd86","Pdcd1lg2","Itga4","Traf1","Ighg3","Cd38")
markers$plasmablast <- c("Bcl6","Havcr2","Cd93","Pdcd1lg2")
markers$plasmatic <- c("Sdc1","Prdm1","Irf4","Xbp1")
markers$regulator <- c("Il12a","Cd24a")
```

```{r}

DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
data.RNA.BCR.filt.wo.dob <- AddModuleScore_UCell(data.RNA.BCR.filt.wo.dob, features = markers)
signature.names <- paste0(names(markers), "_UCell")
VlnPlot_scCustom(data.RNA.BCR.filt.wo.dob,plot_boxplot=TRUE,features = signature.names, group.by ="seurat_clusters")
```

```{r}
y_columns <- c("naive_UCell", "activated_UCell", "prememory_UCell", "memory_UCell", "plasmablast_UCell", "plasmatic_UCell", "regulator_UCell")

# Convert the data to long format
data_long <- data.RNA.BCR.filt.wo.dob_df %>% pivot_longer(cols = all_of(y_columns), names_to = "variable", values_to = "value")

# Create a separate violin plot for each score, grouped by "seurat_clusters"
for (column in y_columns) {
  # Filter the data for the current column
  plot_data <- data_long %>% filter(variable == column)
  
  # Create the ggplot
  p <- ggplot(plot_data, aes(x = seurat_clusters, y = value, fill = seurat_clusters)) +
    geom_violin(width = 1.4) +
    geom_boxplot(width = 0.1, color = "grey", alpha = 0.2) +
    scale_fill_viridis(discrete = TRUE) +
    theme_minimal() +
    theme(
      legend.position = "none",
      plot.title = element_text(size = 11)
    ) +
    ggtitle(paste("Violin Plot of", column, "by Cluster")) +
    xlab("") +
    ylab("Score")
  
  # Print the plot
  print(p)
}
```

```{r}
data.RNA.BCR.filt.wo.dob_df <- as.data.frame(data.RNA.BCR.filt.wo.dob@meta.data)
```


```{r}
FeaturePlot(data.RNA.BCR.filt.wo.dob, reduction = "umap", features = signature.names, ncol = 3,
    order = T)
```

# GSEA analysis

https://crazyhottommy.github.io/scRNA-seq-workshop-Fall-2019/scRNAseq_workshop_3.html

Instalamos Presto.

```{r}
suppressPackageStartupMessages(library(presto))
Bcell.genes <- wilcoxauc(data.RNA.BCR.filt.wo.dob, 'seurat_clusters')
head(Bcell.genes)
```

```{r}
# we have all the genes for each cluster
dplyr::count(Bcell.genes, group)
```

```{r}
library(msigdbr)
library(fgsea)
msigdbr_species()
```

```{r}
print(msigdbr::msigdbr_collections(),n=30)
```

```{r}
# El C7 es immunologic signature gene sets
m_df<- msigdbr(species = "Mus musculus", category = "C7")
head(m_df)
```

Aplicamos un gene set de :https://www.gsea-msigdb.org/gsea/msigdb/cards/GSE11386_NAIVE_VS_MEMORY_BCELL_DN


```{r}
fgsea_sets<- m_df %>% split(x = .$gene_symbol, f = .$gs_name)

fgsea_sets$GSE22886_NAIVE_BCELL_VS_BM_PLASMA_CELL_DN
```

# Clustering

```{r}
Bcell.genes.0 <- Bcell.genes %>% dplyr::filter(group == "0") %>% arrange(desc(logFC),desc(auc)) %>% head(n = 10)
```


```{r}
library(stats)
# select only the feature and auc columns for fgsea, which statistics to use is an open question
# we use the same as the paper
cluster0.genes <- Bcell.genes %>% dplyr::filter(group == "0") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks0<- deframe(cluster0.genes)
cluster1.genes <- Bcell.genes %>% dplyr::filter(group == "1") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks1<- deframe(cluster1.genes)
cluster2.genes <- Bcell.genes %>% dplyr::filter(group == "2") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks2<- deframe(cluster2.genes)
cluster3.genes <- Bcell.genes %>% dplyr::filter(group == "3") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks3<- deframe(cluster3.genes)
cluster4.genes <- Bcell.genes %>% dplyr::filter(group == "4") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks4<- deframe(cluster4.genes)
cluster5.genes <- Bcell.genes %>% dplyr::filter(group == "5") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks5<- deframe(cluster5.genes)
cluster6.genes <- Bcell.genes %>% dplyr::filter(group == "6") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks6<- deframe(cluster6.genes)
cluster7.genes <- Bcell.genes %>% dplyr::filter(group == "7") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks7<- deframe(cluster7.genes)
cluster8.genes <- Bcell.genes %>% dplyr::filter(group == "8") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks8<- deframe(cluster8.genes)
cluster9.genes <- Bcell.genes %>% dplyr::filter(group == "9") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks9<- deframe(cluster9.genes)
cluster10.genes <- Bcell.genes %>% dplyr::filter(group == "10") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks10<- deframe(cluster10.genes)
```

```{r}
fgseaRes0 <- fgsea(fgsea_sets, stats = ranks0, scoreType = "pos",eps=0)
fgseaRes1 <- fgsea(fgsea_sets, stats = ranks1, scoreType = "pos",eps=0)
fgseaRes2 <- fgsea(fgsea_sets, stats = ranks2, scoreType = "pos",eps=0)
fgseaRes3 <- fgsea(fgsea_sets, stats = ranks3, scoreType = "pos",eps=0)
fgseaRes4 <- fgsea(fgsea_sets, stats = ranks4, scoreType = "pos",eps=0)
fgseaRes5 <- fgsea(fgsea_sets, stats = ranks5, scoreType = "pos",eps=0)
fgseaRes6 <- fgsea(fgsea_sets, stats = ranks6, scoreType = "pos",eps=0)
fgseaRes7 <- fgsea(fgsea_sets, stats = ranks7, scoreType = "pos",eps=0)
fgseaRes8 <- fgsea(fgsea_sets, stats = ranks8, scoreType = "pos",eps=0)
fgseaRes9 <- fgsea(fgsea_sets, stats = ranks9, scoreType = "pos",eps=0)
fgseaRes10 <- fgsea(fgsea_sets, stats = ranks10, scoreType = "pos",eps=0)
```

```{r}
fgseaResTidy0 <- fgseaRes0 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy1 <- fgseaRes1 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy2 <- fgseaRes2 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy3 <- fgseaRes3 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy4 <- fgseaRes4 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy5 <- fgseaRes5 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy6 <- fgseaRes6 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy7 <- fgseaRes7 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy8 <- fgseaRes8 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy9 <- fgseaRes9 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy10 <- fgseaRes10 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
```

##  Plot a barplot for with the normalized Enrichment score

## Cluster 0

```{r}
plotEnrichment(fgsea_sets[["GSE13411_NAIVE_BCELL_VS_PLASMA_CELL_UP"]],
               ranks0) + labs(title="GSE13411_NAIVE_BCELL_VS_PLASMA_CELL_UP")
```
	
padj= 4.222452e-07
NES = 4.10810
size= 174

A NES of 4.10810 indicates a substantial enrichment of genes that are upregulated in plasma cells compared to naive B cells.
This value represents the number of genes within the analyzed gene set that are associated with the comparison of interest (in this case, genes upregulated in plasma cells compared to naive B cells).

```{r}
plotEnrichment(fgsea_sets[["GSE17186_MEMORY_VS_NAIVE_BCELL_UP"]],
               ranks0) + labs(title="GSE17186_MEMORY_VS_NAIVE_BCELL_UP")
```
Es importante considerar que la clasificación de los clusters basada únicamente en la expresión génica puede llevar a interpretaciones ambiguas. A veces, los linfocitos B activados pueden expresar genes asociados tanto con la memoria como con la fase ingenua. Por lo tanto, es esencial corroborar estas observaciones con otros marcadores o análisis adicionales para confirmar si el cluster realmente contiene células de memoria o si hay una sobreposición con células ingenuas activadas.

Si los marcadores específicos indican que este cluster tiene tanto células ingenuas como células ingenuas activadas, puede explicar la dispersión de las líneas negras perpendiculares en el análisis GSEA

```{r}
plotEnrichment(fgsea_sets[["GSE25677_MPL_VS_R848_STIM_BCELL_DN"]],
               ranks0) + labs(title="GSE25677_MPL_VS_R848_STIM_BCELL_DN")
```

```{r}
plotEnrichment(fgsea_sets[["GSE12366_PLASMA_CELL_VS_NAIVE_BCELL_DN"]],
               ranks0) + labs(title="GSE12366_PLASMA_CELL_VS_NAIVE_BCELL_DN")
```
Identifica genes que están disminuidos en las células plasmáticas en comparación con los LB naive.
padj= 4.124634e-12
NES = 5.149475
size= 178

# Cluster 1

```{r}
plotEnrichment(fgsea_sets[["GSE42724_NAIVE_BCELL_VS_PLASMABLAST_DN"]],
               ranks1) + labs(title="GSE42724_NAIVE_BCELL_VS_PLASMABLAST_DN")
```
NES= 3.43
padj= 3.24e-06
size= 172

```{r}
plotEnrichment(fgsea_sets[["GSE22886_NAIVE_VS_IGM_MEMORY_BCELL_DN"]],
               ranks1) + labs(title="GSE22886_NAIVE_VS_IGM_MEMORY_BCELL_DN")
```

padj:1.428068e-08
NES: 3.951188
Size: 188

```{r}
plotEnrichment(fgsea_sets[["GSE22886_NAIVE_VS_IGG_IGA_MEMORY_BCELL_DN"]],
               ranks1) + labs(title="GSE22886_NAIVE_VS_IGG_IGA_MEMORY_BCELL_DN")
```

Hay memoria Switched y hay tambien plasmablastos.


# Cluster 2

```{r}
plotEnrichment(fgsea_sets[["GSE13411_PLASMA_CELL_VS_MEMORY_BCELL_DN"]],
               ranks2) + labs(title="GSE13411_PLASMA_CELL_VS_MEMORY_BCELL_DN")
```

padj: 4.28068e-08
NES: 3.98
size:188

```{r}
plotEnrichment(fgsea_sets[["GSE13411_SWITCHED_MEMORY_BCELL_VS_PLASMA_CELL_UP"]],
               ranks2) + labs(title="GSE13411_SWITCHED_MEMORY_BCELL_VS_PLASMA_CELL_UP")
```
```{r}
plotEnrichment(fgsea_sets[["GSE13411_IGM_VS_SWITCHED_MEMORY_BCELL_DN"]],
               ranks2) + labs(title="GSE13411_IGM_VS_SWITCHED_MEMORY_BCELL_DN")
```
Genes down-regulated in comparison of IgM-memory B cells versus Ig isotype switched memory B cells.
NES = 3.75
padj= 9.063686e-07

```{r}
plotEnrichment(fgsea_sets[["GSE22886_IGM_MEMORY_BCELL_VS_BM_PLASMA_CELL_UP"]],
               ranks2) + labs(title="GSE22886_IGM_MEMORY_BCELL_VS_BM_PLASMA_CELL_UP")

```
Genes up-regulated in comparison of memory IgM B cells versus plasma cells from bone marrow and blood.
padj: 3.962311e-17
NES: 6.01

```{r}
plotEnrichment(fgsea_sets[["GSE13411_NAIVE_VS_IGM_MEMORY_BCELL_DN"]],
               ranks2) + labs(title="GSE13411_NAIVE_VS_IGM_MEMORY_BCELL_DN")
```
padj: 3.15507e-07
NES : 3.97

# Cluster 3

```{r}
plotEnrichment(fgsea_sets[["GSE38696_LIGHT_ZONE_VS_DARK_ZONE_BCELL_UP"]],
               ranks3) + labs(title="GSE38696_LIGHT_ZONE_VS_DARK_ZONE_BCELL_UP")
```
Hay expansión clonal e isotype switching. 

```{r}
plotEnrichment(fgsea_sets[["GSE11386_NAIVE_VS_MEMORY_BCELL_DN"]],
               ranks3) + labs(title="GSE11386_NAIVE_VS_MEMORY_BCELL_DN")
```

Hay linfocitos B de memoria.

```{r}
plotEnrichment(fgsea_sets[["GSE13411_IGM_MEMORY_BCELL_VS_PLASMA_CELL_UP"]],
               ranks3) + labs(title="GSE13411_IGM_MEMORY_BCELL_VS_PLASMA_CELL_UP")
GSE22886_IGM_MEMORY_BCELL_VS_BM_PLASMA_CELL_UP
```

```{r}
plotEnrichment(fgsea_sets[["GSE22886_IGM_MEMORY_BCELL_VS_BM_PLASMA_CELL_UP"]],
               ranks3) + labs(title="GSE22886_IGM_MEMORY_BCELL_VS_BM_PLASMA_CELL_UP")
```


```{r}
plotEnrichment(fgsea_sets[["GSE13411_IGM_VS_SWITCHED_MEMORY_BCELL_DN"]],
               ranks3) + labs(title="GSE13411_IGM_VS_SWITCHED_MEMORY_BCELL_DN")
```
Este indica que etán activadas y que tienen un programming temprano hacua linfocitos B memoria.

# Cluster 4

```{r}
plotEnrichment(fgsea_sets[["GSE42724_NAIVE_BCELL_VS_PLASMABLAST_UP"]],
               ranks4) + labs(title="GSE42724_NAIVE_BCELL_VS_PLASMABLAST_UP")
```

```{r}
plotEnrichment(fgsea_sets[["GSE42724_NAIVE_VS_MEMORY_BCELL_DN"]],
               ranks4) + labs(title="GSE42724_NAIVE_VS_MEMORY_BCELL_DN")	
```
Hay memoria.

```{r}
plotEnrichment(fgsea_sets[["GSE42724_NAIVE_VS_B1_BCELL_UP"]],
               ranks4) + labs(title="GSE42724_NAIVE_VS_B1_BCELL_UP")	

```

```{r}
plotEnrichment(fgsea_sets[["GSE42724_MEMORY_BCELL_VS_PLASMABLAST_UP"]],
               ranks4) + labs(title="GSE42724_MEMORY_BCELL_VS_PLASMABLAST_UP")	

```

```{r}
plotEnrichment(fgsea_sets[["GSE42724_NAIVE_BCELL_VS_PLASMABLAST_DN"]],
               ranks4) + labs(title="GSE42724_NAIVE_BCELL_VS_PLASMABLAST_DN")

```


```{r}
plotEnrichment(fgsea_sets[["GSE11961_MEMORY_BCELL_DAY7_VS_GERMINAL_CENTER_BCELL_DAY7_UP"]],
               ranks4) + labs(title="GSE11961_MEMORY_BCELL_DAY7_VS_GERMINAL_CENTER_BCELL_DAY7_UP")
```

No están tan en expansión como el cluster 3.

```{r}
plotEnrichment(fgsea_sets[["GSE22886_NAIVE_VS_IGM_MEMORY_BCELL_DN"]],
               ranks4) + labs(title="GSE22886_NAIVE_VS_IGM_MEMORY_BCELL_DN")
	
```

# Cluster 5

```{r}
plotEnrichment(fgsea_sets[["GSE13411_IGM_VS_SWITCHED_MEMORY_BCELL_DN"]],
               ranks5) + labs(title="GSE13411_IGM_VS_SWITCHED_MEMORY_BCELL_DN")

```

```{r}
plotEnrichment(fgsea_sets[["GSE13411_SWITCHED_MEMORY_BCELL_VS_PLASMA_CELL_UP"]],
               ranks5) + labs(title="GSE13411_SWITCHED_MEMORY_BCELL_VS_PLASMA_CELL_UP")
```

```{r}
plotEnrichment(fgsea_sets[["GSE22886_NAIVE_VS_IGG_IGA_MEMORY_BCELL_DN"]],
               ranks5) + labs(title="GSE22886_NAIVE_VS_IGG_IGA_MEMORY_BCELL_DN")
```
```{r}
plotEnrichment(fgsea_sets[["GSE12845_IGD_POS_BLOOD_VS_NAIVE_TONSIL_BCELL_DN"]],
               ranks5) + labs(title="GSE12845_IGD_POS_BLOOD_VS_NAIVE_TONSIL_BCELL_DN")

```

```{r}
plotEnrichment(fgsea_sets[["GSE13411_IGM_VS_SWITCHED_MEMORY_BCELL_DN"]],
               ranks5) + labs(title="GSE13411_IGM_VS_SWITCHED_MEMORY_BCELL_DN")
```

```{r}
plotEnrichment(fgsea_sets[["GSE28237_EARLY_VS_LATE_GC_BCELL_DN"]],
               ranks5) + labs(title="GSE28237_EARLY_VS_LATE_GC_BCELL_DN")

```

```{r}
plotEnrichment(fgsea_sets[["GSE12366_GC_VS_MEMORY_BCELL_DN"]],
               ranks5) + labs(title="GSE12366_GC_VS_MEMORY_BCELL_DN")

```

Son memoria con switch-classed

# Cluster 6

```{r}
plotEnrichment(fgsea_sets[["GSE42724_MEMORY_BCELL_VS_PLASMABLAST_UP"]],
               ranks6) + labs(title="GSE42724_MEMORY_BCELL_VS_PLASMABLAST_UP")

```

Son LB memoria.

```{r}
plotEnrichment(fgsea_sets[["GSE22886_NAIVE_VS_IGG_IGA_MEMORY_BCELL_DN"]],
               ranks6) + labs(title="GSE22886_NAIVE_VS_IGG_IGA_MEMORY_BCELL_DN")

```


```{r}
plotEnrichment(fgsea_sets[["GSE22886_IGG_IGA_MEMORY_BCELL_VS_BM_PLASMA_CELL_UP"]],
               ranks6) + labs(title="GSE22886_IGG_IGA_MEMORY_BCELL_VS_BM_PLASMA_CELL_UP")
```
Memori< con Tienen isotype switch.

# Cluster 7

```{r}

plotEnrichment(fgsea_sets[["GSE22886_NAIVE_BCELL_VS_NEUTROPHIL_UP"]],
               ranks7) + labs(title="GSE22886_NAIVE_BCELL_VS_NEUTROPHIL_UP")
```
```{r}
plotEnrichment(fgsea_sets[["GSE12366_PLASMA_CELL_VS_NAIVE_BCELL_DN"]],
               ranks7) + labs(title="GSE12366_PLASMA_CELL_VS_NAIVE_BCELL_DN")
```

# CLuster 8

```{r}
plotEnrichment(fgsea_sets[["GSE42724_MEMORY_BCELL_VS_PLASMABLAST_DN"]],
               ranks8) + labs(title="GSE42724_MEMORY_BCELL_VS_PLASMABLAST_DN")

```

```{r}
plotEnrichment(fgsea_sets[["GSE12366_GC_BCELL_VS_PLASMA_CELL_UP"]],
               ranks8) + labs(title="GSE12366_GC_BCELL_VS_PLASMA_CELL_UP")

```

Como que aun no ha llegado a plasmática, es más de centro germinal avanzado.

```{r}
plotEnrichment(fgsea_sets[["GSE38696_LIGHT_ZONE_VS_DARK_ZONE_BCELL_UP"]],
               ranks8) + labs(title="GSE38696_LIGHT_ZONE_VS_DARK_ZONE_BCELL_UP")


```


# Cluster 9

```{r}
#GSE13411_PLASMA_CELL_VS_MEMORY_BCELL_UP
plotEnrichment(fgsea_sets[["GSE12366_PLASMA_CELL_VS_NAIVE_BCELL_UP"]],
               ranks9) + labs(title="GSE12366_PLASMA_CELL_VS_NAIVE_BCELL_UP")
```

Son claramente plasmáticas.

# Cluster 10

```{r}
plotEnrichment(fgsea_sets[["GSE12366_GC_VS_MEMORY_BCELL_UP"]],
               ranks10) + labs(title="GSE12366_GC_VS_MEMORY_BCELL_UP")
```

```{r}
plotEnrichment(fgsea_sets[["GSE11386_NAIVE_VS_MEMORY_BCELL_UP"]],
               ranks10) + labs(title="GSE11386_NAIVE_VS_MEMORY_BCELL_UP")

```

```{r}
plotEnrichment(fgsea_sets[["GSE22886_NAIVE_VS_IGM_MEMORY_BCELL_DN"]],
               ranks10) + labs(title="GSE22886_NAIVE_VS_IGM_MEMORY_BCELL_DN")

```

# Todos los clusters, lo haremos con FindMarkers

El Findmarkers tambien usa suma rango de Wilcoxon

```{r}
Bcell.genes.9 <- FindMarkers(data.RNA.BCR.filt.wo.dob, ident.1 = "9", ident.2 = NULL, logfc.threshold = -Inf, min.pct = -Inf, min.diff.pct = -Inf,test.use = "roc")
head(Bcell.genes.9)
head(Bcell.genes)
```

```{r}
# Naive B cells
Bcell.genes.9 %>% Bcell.genes.9 %>%
  arrange(desc(logFC), desc(auc)) %>%
  head(n = 10)
```

```{r}
library(stats)
# we feel assured to see Cd79a,junb, Fcer2a which are marker genes for naive B cells.

# select only the feature and auc columns for fgsea, which statistics to use is an open question
# we use the same as the paper
cluster9.genes<- Bcell.genes.9 %>%
  arrange(desc(auc)) %>% 
  dplyr::select(rownames(Bcell.genes.9), auc)

ranks<- deframe(cluster9.genes)

head(ranks)
```

```{r}
fgseaRes<- fgsea(fgsea_sets, stats = ranks, scoreType = "pos",eps=0)
```

```{r}
fgseaResTidy <- fgseaRes %>% as_tibble() %>% arrange(desc(NES))


fgseaResTidy %>% 
  dplyr::select(-leadingEdge, -ES) %>% 
  arrange(padj) %>% 
  head()
```

# GSEA analysis 2

https://crazyhottommy.github.io/scRNA-seq-workshop-Fall-2019/scRNAseq_workshop_3.html

Instalamos Presto.

```{r}
suppressPackageStartupMessages(library(presto))
Bcell.genes <- wilcoxauc(data.RNA.BCR.filt.wo.dob, 'seurat_clusters')
```

```{r}
# we have all the genes for each cluster
dplyr::count(Bcell.genes, group)
```

```{r}
library(msigdbr)
library(fgsea)
msigdbr_species()
```

```{r}
print(msigdbr::msigdbr_collections(),n=30)
```

```{r}
# El C7 es immunologic signature gene sets
m_df<- msigdbr(species = "Mus musculus", category = "C7")
```

Aplicamos un gene set de :https://www.gsea-msigdb.org/gsea/msigdb/cards/GSE11386_NAIVE_VS_MEMORY_BCELL_DN

Las que aplicamos antes:

```{r}
fgsea_sets<- m_df %>% split(x = .$gene_symbol, f = .$gs_name)

markers_plasm <- list()
markers_plasm$plasm <- list(fgsea_sets$GSE22886_NAIVE_BCELL_VS_BM_PLASMA_CELL_DN)
```

```{r}
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
data.RNA.BCR.filt.wo.dob <- AddModuleScore_UCell(data.RNA.BCR.filt.wo.dob, features = markers_plasm)
signature.names <- paste0(names(markers_plasm), "_UCell")
VlnPlot_scCustom(data.RNA.BCR.filt.wo.dob,plot_boxplot=TRUE,features = signature.names, group.by ="seurat_clusters")
```

```{r}
markers <- list()
markers$naive <- list(fgsea_sets$GSE12366_PLASMA_CELL_VS_NAIVE_BCELL_DN)
markers$activadas <- list(fgsea_sets$GSE13547_2H_VS_12_H_ANTI_IGM_STIM_BCELL_UP)
markers$memory_IGM <- list(fgsea_sets$GSE22886_IGM_MEMORY_BCELL_VS_BM_PLASMA_CELL_UP)
markers$memory_IGG_IGA <-list(fgsea_sets$GSE22886_NAIVE_VS_IGG_IGA_MEMORY_BCELL_DN)
markers$plasmablasto <- list(fgsea_sets$GSE42724_MEMORY_BCELL_VS_PLASMABLAST_DN)
markers$plasmatic <- list(fgsea_sets$GSE12366_PLASMA_CELL_VS_NAIVE_BCELL_UP)
markers$GC <- list(fgsea_sets$GSE12366_GC_VS_NAIVE_BCELL_UP)
markers$switched <- list(fgsea_sets$GSE13411_SWITCHED_MEMORY_BCELL_VS_PLASMA_CELL_UP)
markers$high_affinity <- list(c("Batf+", "Cd320+", "Aurka+", "Nfil3+", "Pml+", "Bach2-", "Egr3-", "Elk4-", "Foxp1-", "Jun-", "Nr4a1-", "Rel-", "Bub1+", "Ccna2+", "Ccnb1+", "Ccnb2+", "Ccnd2+", "Cdc20+", "Cdc25c+", "Kif22+", "Plk1+", "Timd2+", "Tnfrsf8-", "Ccr6-", "Cd22-", "Cd38-", "Cd72-", "Cr2-", "Fcer2a-", "Icosl-", "Pecam1-", "Siglecg-", "Tlrl-", "Tnfrsf18-", "Gars+", "Gart+", "ler3+", "Mif+", "Myc+", "Spp1+", "Uck2+", "Ppp1r15a-", "Rgs1-"))
```

```{r}
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
data.RNA.BCR.filt.wo.dob <- AddModuleScore_UCell(data.RNA.BCR.filt.wo.dob, features = markers)
signature.names <- paste0(names(markers), "_UCell")
VlnPlot_scCustom(data.RNA.BCR.filt.wo.dob,plot_boxplot=TRUE,features = signature.names, group.by ="seurat_clusters")
```

```{r}
VlnPlot(object = data.RNA.BCR.filt.wo.dob, features = 'Bach2')
```


# SignatuR

```{r}
SignatuR
```

Me interesa:
- Immunoglobulins
- cell_types

```{r}
print(SignatuR, "Reference","Signature")
```

```{r}
library(DiagrammeR)
plot(SignatuR)
```


```{r}
markers_cc <- list()

cycling.G1S <- GetSignature(SignatuR$Mm$Programs$cellCycle.G1S)
cycling.G2M <- GetSignature(SignatuR$Mm$Programs$cellCycle.G2M)

markers_cc$G1S <- GetSignature(SignatuR$Mm$Programs$cellCycle.G1S)
markers_cc$G2M <- GetSignature(SignatuR$Mm$Programs$cellCycle.G2M)
```

```{r}
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
data.RNA.BCR.filt.wo.dob <- AddModuleScore_UCell(data.RNA.BCR.filt.wo.dob, features = markers_cc)
signature.names <- paste0(names(markers_cc), "_UCell")
VlnPlot_scCustom(data.RNA.BCR.filt.wo.dob,plot_boxplot=TRUE,features = signature.names, group.by ="seurat_clusters")
```

```{r}
FeaturePlot(data.RNA.BCR.filt.wo.dob, reduction = "umap", features = signature.names, ncol = 3,
    order = T)
```

```{r}
#esto a mano, te lo pasa a minuscula, funciona muy bien
m.s.genes <- str_to_title(cc.genes.updated.2019$s.genes)
m.g2m.genes <- str_to_title(cc.genes.updated.2019$g2m.genes)
# #https://github.com/satijalab/seurat/issues/4617
# m.s.genes
# m.g2m.genes
DefaultAssay(data.RNA.BCR.filt.wo.dob) <- "RNA"
markers_cc_seurat <- list()
markers_cc_seurat$s <- list(m.s.genes)
markers_cc_seurat$g2m <- list(m.g2m.genes)

data.RNA.BCR.filt.wo.dob <- AddModuleScore_UCell(data.RNA.BCR.filt.wo.dob, features = markers_cc_seurat)
signature.names <- paste0(names(markers_cc_seurat), "_UCell")
FeaturePlot(data.RNA.BCR.filt.wo.dob, reduction = "umap", features = signature.names, ncol = 2,
    order = T)
```

```{r}
LB.scores <- ScoreSignatures_UCell(data.RNA.BCR.filt.wo.dob@assays$RNA@data, features = markers)
head(LB.scores)
```

```{r}
library(reshape2)
library(ggplot2)
melted <- reshape2::melt(LB.scores)
colnames(melted) <- c("Cell", "Signature", "UCell_score")
p <- ggplot(melted, aes(x = Signature, y = UCell_score)) + geom_violin(aes(fill = Signature),
    scale = "width") + geom_boxplot(width = 0.1, outlier.size = 0) + theme_bw() +
    theme(axis.text.x = element_blank())
p
```

```{r}
set.seed(123)
ranks <- StoreRankings_UCell(data.RNA.BCR.filt.wo.dob@assays$RNA@data)
ranks[1:5, 1:5]
```

```{r}
set.seed(123)
u.scores.2 <- ScoreSignatures_UCell(features = signatures, precalc.ranks = ranks)

melted <- reshape2::melt(u.scores.2)
colnames(melted) <- c("Cell", "Signature", "UCell_score")
p <- ggplot(melted, aes(x = Signature, y = UCell_score)) + geom_violin(aes(fill = Signature),
    scale = "width") + geom_boxplot(width = 0.1, outlier.size = 0) + theme_bw() +
    theme(axis.text.x = element_blank())
p
```


```{r}
Bcell.genes.0 <- Bcell.genes %>% dplyr::filter(group == "0") %>% arrange(desc(logFC),desc(auc)) %>% head(n = 10)
```

# esto de un tutorial usando singleseqgset

https://arc85.github.io/singleseqgset/articles/singleseqgset.html

```{r}
m_df<- msigdbr(species = "Mus musculus", category = "C7")
m.names <- unique(m_df$gs_name)

m.sets <- vector("list",length=length(m.names))
names(m.sets) <- m.names

for (i in names(m.sets)) {
    m.sets[[i]] <- pull(m_df[m_df$gs_name==i,"gene_symbol"])
}

```

```{r}
library(singleseqgset)
logfc.data <- logFC(cluster.ids = data.RNA.BCR.filt.wo.dob@meta.data$seurat_clusters, expr.mat = data.RNA.BCR.filt.wo.dob@assays$SCT@data)
names(logfc.data)
```

```{r}
gse.res <- wmw_gsea(expr.mat=data.RNA.BCR.filt.wo.dob@assays$SCT@data,cluster.cells=logfc.data[[1]],log.fc.cluster=logfc.data[[2]],gene.sets=m.sets)
```


```{r}
library(stats)
# select only the feature and auc columns for fgsea, which statistics to use is an open question
# we use the same as the paper
cluster0.genes <- Bcell.genes %>% dplyr::filter(group == "0") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks0<- deframe(cluster0.genes)
cluster1.genes <- Bcell.genes %>% dplyr::filter(group == "1") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks1<- deframe(cluster1.genes)
cluster2.genes <- Bcell.genes %>% dplyr::filter(group == "2") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks2<- deframe(cluster2.genes)
cluster3.genes <- Bcell.genes %>% dplyr::filter(group == "3") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks3<- deframe(cluster3.genes)
cluster4.genes <- Bcell.genes %>% dplyr::filter(group == "4") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks4<- deframe(cluster4.genes)
cluster5.genes <- Bcell.genes %>% dplyr::filter(group == "5") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks5<- deframe(cluster5.genes)
cluster6.genes <- Bcell.genes %>% dplyr::filter(group == "6") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks6<- deframe(cluster6.genes)
cluster7.genes <- Bcell.genes %>% dplyr::filter(group == "7") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks7<- deframe(cluster7.genes)
cluster8.genes <- Bcell.genes %>% dplyr::filter(group == "8") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks8<- deframe(cluster8.genes)
cluster9.genes <- Bcell.genes %>% dplyr::filter(group == "9") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks9<- deframe(cluster9.genes)
cluster10.genes <- Bcell.genes %>% dplyr::filter(group == "10") %>%   arrange(desc(auc)) %>% dplyr::select(feature, auc)
ranks10<- deframe(cluster10.genes)
```

```{r}
fgseaRes0 <- fgsea(fgsea_sets, stats = ranks0, scoreType = "pos",eps=0)
fgseaRes1 <- fgsea(fgsea_sets, stats = ranks1, scoreType = "pos",eps=0)
fgseaRes2 <- fgsea(fgsea_sets, stats = ranks2, scoreType = "pos",eps=0)
fgseaRes3 <- fgsea(fgsea_sets, stats = ranks3, scoreType = "pos",eps=0)
fgseaRes4 <- fgsea(fgsea_sets, stats = ranks4, scoreType = "pos",eps=0)
fgseaRes5 <- fgsea(fgsea_sets, stats = ranks5, scoreType = "pos",eps=0)
fgseaRes6 <- fgsea(fgsea_sets, stats = ranks6, scoreType = "pos",eps=0)
fgseaRes7 <- fgsea(fgsea_sets, stats = ranks7, scoreType = "pos",eps=0)
fgseaRes8 <- fgsea(fgsea_sets, stats = ranks8, scoreType = "pos",eps=0)
fgseaRes9 <- fgsea(fgsea_sets, stats = ranks9, scoreType = "pos",eps=0)
fgseaRes10 <- fgsea(fgsea_sets, stats = ranks10, scoreType = "pos",eps=0)
```

```{r}
fgseaResTidy0 <- fgseaRes0 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy1 <- fgseaRes1 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy2 <- fgseaRes2 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy3 <- fgseaRes3 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy4 <- fgseaRes4 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy5 <- fgseaRes5 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy6 <- fgseaRes6 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy7 <- fgseaRes7 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy8 <- fgseaRes8 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy9 <- fgseaRes9 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
fgseaResTidy10 <- fgseaRes10 %>% as_tibble() %>% arrange(desc(NES)) %>%  dplyr::select(-leadingEdge, -ES) %>%  arrange(padj)
```
